requirements.txt
# Core dependencies
apache-airflow==2.7.1
pyspark==3.5.0

# Data processing
pandas==2.1.1
pyarrow==13.0.0
duckdb==0.9.1

# NLP
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.1

# Graph
networkx==3.2

# Vector store
chromadb==0.4.15

# Web scraping
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2

# Visualization
streamlit==1.28.0
plotly==5.17.0
folium==0.15.0

# Utilities
python-dotenv==1.0.0

# Gemini AI and web scraping
google-generativeai==0.8.5
beautifulsoup4==4.12.2